backend/requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
playwright
python-dotenv
beautifulsoup4
requests